Move BeaconBlockBodyLists to testFixtures
