+++ ethereum/datastructures/src/testFixtures/java/tech/pegasys/teku/datastructures/util/BeaconBlockBodyLists.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.datastructures.blocks;
+package tech.pegasys.teku.datastructures.util;
 
 import tech.pegasys.teku.datastructures.operations.Attestation;
 import tech.pegasys.teku.datastructures.operations.AttesterSlashing;
